+++ Lib/I2C/I2C.h
@@ -12,7 +12,7 @@
 void I2C_Init(void);
 uint8_t I2C_Scan(void);
 void Send_I2C_Data(uint8_t databyte);
-unsigned int Read_I2C_Data(void);
+uint8_t Read_I2C_Data(void);
 void Send_I2C_ControlByte(uint8_t Dev_Add,uint8_t RW_bit);
 void Send_I2C_StartBit(void);
 void Send_I2C_StopBit(void);
